Fix off by one error in sub_string tests
strlen returns the length of the string excluding the null terminator,
so we have to add 1 to ensure that the length of the output buffer is
long enough. This only matters if the sub string is the same length as
the string (i.e. the whole string is returned), but this is a valid use
case as the string is a sub string of itself.